requirements.txt
python-dotenv
psycopg2-binary
clickhouse-connect
apache-airflow
mimesis